+++ CHANGES.md
@@ -59,6 +59,7 @@ Changes in 13.0
 - Add RosarioSIS\Functions\AddonInfo class in AddonInfo.php
 - Check add-on requirements in Modules.inc.php & Plugins.inc.php
 - Add AddonMakeUpdateAvailable() function in Addon.fnc.php
+- Check for add-on updates in Modules.inc.php & Plugins.inc.php
 
 Changes in 12.9.2
 -----------------